formatage de type de données
